requirements.txt
Django==1.9.2
Markdown==2.6.5
argparse==1.2.1
django-filter==0.12.0
djangorestframework==3.3.2
wsgiref==0.1.2
